若存在推荐失败的情况，清空redis中相关简历数据,否则不清楚

+++ profile-service-provider/src/main/java/com/moseeker/profile/service/impl/ReferralServiceImpl.java
@@ -553,10 +553,9 @@ public class ReferralServiceImpl implements ReferralService {
                 .buildEmployeeReferralProfileNotice();
         List<MobotReferralResultVO> referralResultVOS = referralProfileFileUpload.employeeReferralProfileAdaptor(profileNotice);
         List<Integer> referralIds = referralResultVOS.stream().map(MobotReferralResultVO::getId).collect(Collectors.toList());
-        client.del(AppId.APPID_ALPHADOG.getValue(), KeyIdentifier.EMPLOYEE_REFERRAL_PROFILE.toString(), String.valueOf(employeeId));
-//        if(com.moseeker.common.util.StringUtils.isEmptyList(referralIds)){
-//            throw CommonException.PROGRAM_EXCEPTION;
-//        }
+        if(!com.moseeker.common.util.StringUtils.isEmptyList(referralIds)){//若存在推荐失败的情况，清空redis中相关简历数据
+            client.del(AppId.APPID_ALPHADOG.getValue(), KeyIdentifier.EMPLOYEE_REFERRAL_PROFILE.toString(), String.valueOf(employeeId));
+        }
 
         referralResultVOS = referralResultVOS.stream().map(resultVO -> {
             //获取该职位应得的积分
